refactor processing controllers in dedicated package
