+++ Portal/vip-api/src/main/java/fr/insalyon/creatis/vip/api/rest/controller/processing/ExecutionControler.java
@@ -29,7 +29,7 @@
  * The fact that you are presently reading this means that you have had
  * knowledge of the CeCILL-B license and that you accept its terms.
  */
-package fr.insalyon.creatis.vip.api.rest.controller;
+package fr.insalyon.creatis.vip.api.rest.controller.processing;
 
 import fr.insalyon.creatis.vip.api.bean.Execution;
 import fr.insalyon.creatis.vip.api.business.*;
@@ -48,10 +48,7 @@ import org.springframework.web.bind.annotation.*;
 import javax.servlet.http.HttpServletRequest;
 import javax.validation.Valid;
 
-import java.util.List;
-
 import static fr.insalyon.creatis.vip.api.CarminProperties.DEFAULT_LIMIT_LIST_EXECUTION;
-import static jdk.nashorn.internal.runtime.regexp.joni.constants.StackType.POS;
 
 /**
  * Created by abonnet on 7/13/16.
